refactor `state` `smt`

+++ apps/registry/src/lib.rs
@@ -1,7 +1,7 @@
 use odana_rt_api::Executable;
 use template::{WASM_BINARY, metadata};
 
-pub fn built_in_apps() -> Vec<(u64, Executable)> {
+pub fn built_in_apps() -> Vec<(u32, Executable)> {
     let mut apps = Vec::new();
     if let Some(template_wasm_binary) = WASM_BINARY {
         apps.push((0, Executable {
